% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleBarcharts}
\alias{exampleBarcharts}
\title{exampleBarcharts}
\usage{
exampleBarcharts()
}
\value{
Opens a  Shiny app
}
\description{
Show a Plotly and GGPlot2 implementation of a simple barchart
}
\examples{
\dontrun{exampleBarcharts()}
}
