% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleBenchmarkGroupMeans}
\alias{exampleBenchmarkGroupMeans}
\title{exampleBenchmarkGroupMeans}
\usage{
exampleBenchmarkGroupMeans(type = c("run", "view"))
}
\value{
Show a Plotly plot of the benchmark'
}
\description{
This example shows the result of a benchmark between dplyr and data.table on comparing group means.
See: examples/dh/dh.benchmark.groupmeans.R for a more detailed description
}
\examples{
\dontrun{exampleBenchmarkGroupMeans()}
}
