% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleBenchmarkSubsetting}
\alias{exampleBenchmarkSubsetting}
\title{exampleBenchmarkSubsetting}
\usage{
exampleBenchmarkSubsetting(type = c("run", "view"))
}
\value{
Show a Plotly plot of the benchmark'
}
\description{
This example shows the result of a benchmark between dplyr and data.table on subsetting a large data set.
See: examples/dh/dh.benchmark.subsetting.R for a more detailed description
}
\examples{
\dontrun{exampleBenchmarkSubsetting()}
}
