% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleShinyCatMoneyHappy}
\alias{exampleShinyCatMoneyHappy}
\title{exampleShinyCatMoneyHappy}
\usage{
exampleShinyCatMoneyHappy(type = c("run", "view"))
}
\arguments{
\item{type}{"run" or "view"}
}
\description{
This example shows
}
\examples{
\dontrun{exampleShinyCatMoneyHappy()}
}
