% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIE2024-package.R
\docType{package}
\name{MIE2024-package}
\alias{MIE2024}
\alias{MIE2024-package}
\title{MIE2024: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
